pyproject.toml
[project]
name = "itmo-tg-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Victor",email = "69643322+blueberry13-8@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "chromadb (>=1.0.17,<2.0.0)",
    "tiktoken (>=0.11.0,<0.12.0)",
    "aiogram (>=3.21.0,<4.0.0)",
    "langchain-chroma (>=0.2.5,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
